fix: keybindings for shortcut
